+++ packages/nextclaw-remote/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nextclaw/remote
 
+## 0.1.21
+
+### Patch Changes
+
+- Updated dependencies
+  - @nextclaw/server@0.10.25
+
 ## 0.1.20
 
 ### Patch Changes